pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quantum-time-travel"
version = "0.1.0"
description = "FSM- and graph-based planning core with quantum backend adapters (stubs)"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
  { name = "Idyll Intelligent Systems" },
]
keywords = ["fsm", "graph", "planner", "quantum", "time-travel", "simulation"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = []

[project.optional-dependencies]
qiskit = ["qiskit>=1.2.0"]
cirq = ["cirq>=1.3.0"]
braket = ["amazon-braket-sdk>=1.74.0"]
api = ["fastapi>=0.110", "uvicorn[standard]>=0.29"]
dev = ["pytest>=7.4", "ruff>=0.5.0", "mypy>=1.7.0", "fastapi>=0.110", "uvicorn[standard]>=0.29"]

[project.scripts]
quantum-time-travel = "quantum_time_travel.cli:main"
quantum-time-travel-api = "uvicorn:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
strict = false
ignore_missing_imports = true
